requirements.txt
torch
transformers
deepspeed
git+https://github.com/huggingface/peft.git
colorlog
requests
pillow
numpy
google-cloud-storage
